Disable binary upgrades, Update tooling
Since Go 1.9, the list of releases are too long
for the Go team to list on a single page (what?).
Therefore, we have to disable upgrades since I was
previously parsing it all. Future changes could
be to hust hardcode pulling from "page 2"; or, tell
the Go team to list it all on a single page.  Blah.

Updating tooling:

- all x/tools/cmd/...
- disable go-find-references, since it is still broken
- add go-langserver (mostly for vscode)
- add Hugo tooling now