Increase git speeds by 400%
This combined 4 git commands into 1 to increase
the speed of git drastically.

Now using 'git status --porcelain --branch --show-stash'
and parsing it over and over again to increase speed.

NOTE: There is a bug in git that --porcelain does not
return the Stash count with --show-stash.  I have filed
a bug with the Git team.  In the meantime, I have hardcoded
the git stash list, which is very very slow.

A future version will hopefully fix git status --porcelain
to include Stash count, and therefore we can remove the
extra "git stash list" command.

But until then, you may want to disable the stash parts
if it gets too slow.